feat: setup different emailing services for dev and test
