+++ src/main/resources/application.properties
@@ -13,6 +13,7 @@ spring.jpa.properties.hibernate.format_sql=true
 # logging.level.org.hibernate=INFO
 
 # Spring Boot Configuration
+spring.profiles.active=dev
 spring.application.name="Ensak Connect"
 server.port=8081
 server.error.include-message=always
