+++ src/test/resources/application.properties
@@ -1,3 +1,5 @@
+spring.profiles.active=test
+
 spring.datasource.url=jdbc:h2:mem:db;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
